% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{GPActivity}
\alias{GPActivity}
\alias{GPActivity-package}
\alias{gpa.process}
\title{GPActivity}
\usage{
gpa.process(epoch.file, frequency)
}
\description{
This package contains functions that accept a 60sec epoch .csv file,
and generates summaries of sleep time/efficiency, non-wear time, and MET/min.
}

